+++ src/control_plane/shared_infrastructure/postgres/mod.rs
@@ -10,6 +10,7 @@ mod prepare_postgres_shared_instances;
 mod prepared_postgres_shared_instance;
 mod provision_postgres_logical_resource;
 mod reconcile_postgres_project_resources;
+mod reconcile_prepared_postgres_instance;
 
 pub(super) const POSTGRES_BOOTSTRAP_USERNAME: &str = "stackctl_admin";
 
@@ -25,3 +26,4 @@ pub(crate) use prepare_postgres_shared_instances::prepare_postgres_shared_instan
 pub(crate) use prepared_postgres_shared_instance::PreparedPostgresSharedInstance;
 pub(crate) use provision_postgres_logical_resource::provision_postgres_logical_resource;
 pub(crate) use reconcile_postgres_project_resources::reconcile_postgres_project_resources;
+pub(crate) use reconcile_prepared_postgres_instance::reconcile_prepared_postgres_instance;